Influfinder/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-swagger-ui==4.11.1
python-dotenv==1.0.1
prometheus-client==0.20.0
openpyxl==3.1.2
python-json-logger==2.0.7
requests==2.32.3
pytest==8.2.2
